pyproject.toml
[project]
name = "daily-news-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "crewai[tools]>=0.152.0",
    "firecrawl-py>=4.3.5",
    "python-dotenv>=1.1.1",
]
